refactor: refactors marquee selection into a custom hook
Moves marquee selection logic into a reusable custom hook to improve code organization and maintainability.

This hook handles the selection box drawing, mouse events, and item selection within the defined area.
