+++ src/components/Builder/Scene.js
@@ -2,10 +2,8 @@
 import {
   createRef,
   Fragment,
-  useCallback,
   useEffect,
   useRef,
-  useState,
 } from 'react';
 import * as classNames from '../../constants/classNames';
 import ContextMenu from './ContextMenu';
@@ -20,33 +18,24 @@ import {
   calculateGuidePositions,
   findItemById,
   findItemsOnPage,
-  getItemsInSelectionBox,
 } from '../../utils/functions';
 import DraggableLayer from './DraggableLayer';
 import useKeyboardActions from '../../utils/useKeyboardActions';
+import useMarqueeSelection from '../../utils/useMarqueeSelection';
 
 const Scene = () => {
   const pages = usePropStore(state => state.pages);
   const settings = usePropStore(state => state.settings);
   const contextMenuProps = useBuilderStore(state => state.contextMenuProps);
-
-  const resetActiveElements = useBuilderStore(state => state.resetActiveElements);
-  const setActiveElementsSelection = useBuilderStore(state => state.setActiveElementsSelection);
   const setContextMenuProps = useBuilderStore(state => state.setContextMenuProps);
   const setGuides = useBuilderStore(state => state.setGuides);
-
   const zoom = useBuilderStore(state => state.zoom);
-
-  // Marquee selection state
-  const [isSelecting, setIsSelecting] = useState(false);
-  const [selectionBox, setSelectionBox] = useState(null);
-  const [selectionPageId, setSelectionPageId] = useState(null);
-  const selectionStartRef = useRef(null);
   const lastScrollPosition = useBuilderStore(state => state.lastScrollPosition);
 
   const pageStyles = useRef({});
   const pageContainerStyles = useRef({});
   const viewPortRef = useRef({});
+  const canvasRef = useRef(null);
 
   /* Page Refs */
   const refs = useRef(pages.reduce((acc, curr) => {
@@ -54,8 +43,9 @@ const Scene = () => {
     return acc;
   }, {}));
 
-  // Handle keyboard actions
+  // Custom hooks
   useKeyboardActions();
+  const { selectionBox, selectionPageId } = useMarqueeSelection(canvasRef);
 
   // Update refs when new pages are added
   useEffect(() => {
@@ -66,6 +56,7 @@ const Scene = () => {
     });
   }, [pages]);
 
+  // Calculate guides for snap functionality
   useEffect(() => {
     setGuides(pages.reduce((acc, page) => {
       const _pageGuides = {};
@@ -93,121 +84,12 @@ const Scene = () => {
     }, {}));
   }, [pages, zoom, setGuides]);
 
+  // Restore scroll position after mode change
   useEffect(() => {
     if (viewPortRef.current) {
       viewPortRef.current.scrollTop = lastScrollPosition;
     }
-  }, [lastScrollPosition]); // set last scroll position after changing mode
-
-  // Marquee selection handlers
-  const handleCanvasMouseDown = useCallback(e => {
-    // Selection starts with left click only
-    if (e.button !== 0) return;
-
-    const { target } = e;
-
-    // Don't start selection if clicking on an item or interactive element
-    const activeElementsSelectors = [
-      '.reportItem',
-      '.reportItemWrapper',
-      '.pageActions',
-      '.jfReport-addSlide',
-      '[role="button"]',
-      'button',
-      '.contextMenu',
-    ];
-
-    if (activeElementsSelectors.some(selector => target.closest(selector))) {
-      return;
-    }
-
-    // Find the page element from the click target
-    const pageElement = target.closest('.jfReport-page') || target.closest('[id*="presentation-page-"]');
-    if (!pageElement) return;
-
-    const pageId = pageElement.getAttribute('data-id') || pageElement.getAttribute('id').replace('presentation-page-', '');
-    if (!pageId) return;
-
-    // Prevent text selection during drag
-    e.preventDefault();
-
-    const pageRect = pageElement.getBoundingClientRect();
-    const startX = e.clientX - pageRect.left;
-    const startY = e.clientY - pageRect.top;
-
-    selectionStartRef.current = {
-      pageRect,
-      startX,
-      startY,
-    };
-    setSelectionPageId(pageId);
-    setSelectionBox({
-      endX: startX,
-      endY: startY,
-      startX,
-      startY,
-    });
-    setIsSelecting(true);
-
-    // Clear current selection when starting new marquee
-    resetActiveElements();
-  }, [resetActiveElements]);
-
-  const handleCanvasMouseMove = useCallback(e => {
-    if (!isSelecting || !selectionStartRef.current) return;
-
-    const { pageRect, startX, startY } = selectionStartRef.current;
-
-    // Calculate end position (clamped to page bounds)
-    const endX = Math.max(0, Math.min(e.clientX - pageRect.left, pageRect.width));
-    const endY = Math.max(0, Math.min(e.clientY - pageRect.top, pageRect.height));
-
-    setSelectionBox({
-      endX,
-      endY,
-      startX,
-      startY,
-    });
-  }, [isSelecting]);
-
-  const handleCanvasMouseUp = useCallback(() => {
-    if (!isSelecting || !selectionBox || !selectionPageId) {
-      setIsSelecting(false);
-      setSelectionBox(null);
-      selectionStartRef.current = null;
-      return;
-    }
-
-    // Find the page and its items
-    const page = pages.find(p => p.id === selectionPageId);
-
-    if (page && page.items) {
-      const selectedItemIds = getItemsInSelectionBox(selectionBox, page.items, zoom);
-      if (selectedItemIds.length > 0) {
-        setActiveElementsSelection(selectedItemIds);
-      }
-    }
-
-    // Reset selection state
-    setIsSelecting(false);
-    setSelectionBox(null);
-    setSelectionPageId(null);
-    selectionStartRef.current = null;
-  }, [isSelecting, selectionBox, selectionPageId, pages, zoom, setActiveElementsSelection]);
-
-  // Add global mouse event listeners for marquee selection
-  useEffect(() => {
-    if (isSelecting) {
-      document.body.style.cursor = 'crosshair';
-      document.addEventListener('mousemove', handleCanvasMouseMove);
-      document.addEventListener('mouseup', handleCanvasMouseUp);
-      return () => {
-        document.body.style.cursor = '';
-        document.removeEventListener('mousemove', handleCanvasMouseMove);
-        document.removeEventListener('mouseup', handleCanvasMouseUp);
-      };
-    }
-  }, [isSelecting, handleCanvasMouseMove, handleCanvasMouseUp]);
+  }, [lastScrollPosition]);
 
   const { reportLayoutHeight = 794, reportLayoutWidth = 1123 } = settings;
 
@@ -242,10 +124,8 @@ const Scene = () => {
         data-zoom={zoom}
       >
         <div
+          ref={canvasRef}
           className={classNames.canvas}
-          onMouseDown={handleCanvasMouseDown}
-          onMouseMove={handleCanvasMouseMove}
-          onMouseUp={handleCanvasMouseUp}
         >
           {pages.map((page, index) => (
             // TODO: This part can be moved into a different component
@@ -269,11 +149,8 @@ const Scene = () => {
                   pageIndex={index}
                   style={pageContainerStyles.current}
                 />
-                {selectionPageId === page.id && (
-                  <SelectionBox
-                    isSelecting={isSelecting}
-                    selectionBox={selectionBox}
-                  />
+                {selectionPageId === page.id && selectionBox && (
+                  <SelectionBox selectionBox={selectionBox} />
                 )}
               </div>
             </Fragment>
